added a chest in the spawn room along the right wall. added ClearHeroInteractionRec() to InteractionFunctions. add codepath to Handle() Hero & Chest interaction, sets chest to empty chest. updated ActorFunctions.Update() to only lock actor if an interaction should actually happen.
AND GUESS WHAT? we don't even need to press the A button to open the chest!
the hero's collision component rec overlaps the chest for a moment, and the interaction (Handle()) happens.
BUT we will need the interaction rec for when hero wants to pick up a pot or liftable object.
we'll also need it when hero wants to drag/push/pull and object.