added reward value to actor.state enum. added reward animation group to actor animation list, pointed it to use animation. updated actorFunctions.Update() with codepath to handle state.Reward. this state duplicates the state of use. moved 50 gold reward sprite into particles/projectile sheet, out of dungeon sheet. added codepath for 50 gold reward to GameObjectFunctions.SetType().
added call to spawn a reward particle over the actor that opened the chest.
extended the lifetime of the reward particle so player has longer to 'read' it.

playtested, noted some changes i'd like to make.
moved the reward particle to spawn over the chest instead of the player.

playtested, reward state not locking upon chest interaction.
removed codepath from ActorFunctions.Update() for reward state.
expanded code for setting reward state in interactionFunctions.Handle().
playtested, nah fam it's still not working. reward animation doesn't display.

cause: there isn't a codepath for setting the animation group in SetAnimationGroup().
fixed it by adding in that codepath to SetAnimationGroup().
cleaned up the implementation of reward state across various classes (the mess created above).
finessed the placement of the reward particle to be over the actor's head.