updated frames for hearts + item/weapon background sprites in WorldUI class.
updated SpawnProjectile() to accept a Direction parameter.
this lets us set a projectile's direction upon spawn, instead of always inheriting actor's direction.
updated UseWeapon() to pass Actor.Direction as parameter to SpawnProjectile().

updated SpawnProjectile() to accept a Vector2 Position parameter instead of Actor parameter.
this removes the dependence on an actor, and allows us to spawn projectiles without an actor reference.
this will cause a problem later tho, when we want to spawn a boomerang projectile.
the boomerang projectile will want to return to the actor that spawned it.
we could limit the boomerang projectile to only the hero, then track the hero's position via Pool.hero.
but that means other actors could never spawn a boomerang projectile.

playtested, working.