updated Input to store the last frame's gamePadDirection in an instance called LastGamePadDirection. created MenuWidgetInventory.cs, filled with basic methods taken from MenuWidgetInfo. added calls to MenuWidgetInventory from InventoryScreen. playtested, working as expected.
added a list of dividers to MenuWidgetInventory.
added a list of text components to MenuWidgetInventory, to hold the 'weapons, armor, equipment' labels.
added a list of menuItems to MenuWidgetInventory, to hold all the menuItems.
updated Reset() to place and align dividers, labels, and menuItems.
created PlaceRow(), and PlaceMenuItem() to simplify the placement of menuItems in InventoryWidget.

created MenuItemFunctions.cs.
created SetNeighbors(List<MenuItem>, int RowLength).
calling SetNeighbors() from InventoryWidget.
SetNeighbors() is a beautiful function that sets the neighbors of menuItems based on a rowLength parameter.
this function saves me HUNDREDS of lines of code, and work generically on any list of menuItems.

added selectionBox component sprite to InventoryScreen.
updated InventoryScreen.HandleInput() to move the selectionBox between menuItems for the InventoryWidget.