cleaned up the outline heart sprite, removed unneeded pixels. reduced opacity. rearranged how the heart piece sprites were laid out in the main sheet.
updated MenuWidgetLoadout.Reset() to set the heart piece sprite based on WorldUI.pieceCounter.
this pieceCounter is updated each frame, so it will always have the proper value.
best of all, this is done without using the modulo (%) operator, which is a very expensive operation.
instead it's done using addition and two counters, in WorldUI.Update().
playtested, working.