doubled up on the explosion's first frame, so the initial impact reads better. playtested, it's much more satisfying when you destroy an enemy now. reduced the sorting offset for fire particle just a bit, to sort better with actors.
changed how the reward animation list for actors is defined.
it no longer points to the use animation, it defines it's own frames to use.
the actor.use animation now points to the attack animation.
this fixes a bug where hero was set into the 'reward' animation upon a use state.
playtested, looking MUCH better and natural.