optimization: don't move, collision check, or handle interactions for non-blocking room objects. these objects will never collide/interact with other objects. other objects WILL collide & interact with them. but there's no need to do this checking/interaction twice.
change: added a check into PoolFunctions.Move() to only Move() non-blocking objects in the objPool.
this dropped the Update loop from around 2200 ticks to around 600 ticks (for spawn room, sans mobs).
so, a savings of about 1600 ticks, or a reduction in time of 73%.
an optimization that reduces each frame's total time by 73% isn't bad.