created gold UI animated sprites in MainSheet. updated menuItem to have animation component. updated SetMenuItemData() to set menuItem.compAnim.currentAnimation lists.
updated MenuWidgetInfo to use a MenuItem instead of a component sprite.
this is because MenuItems animate now, so it's just easier to use a menuItem here, instead of two components.

updated MenuWidgetInfo to animate it's infoItem.
updated MenuWidgetLoadout to animate it's menuItems.
updated MenuWidgetInventory to animate it's menuItems.
updated MenuWidgetOptions to animate it's menuItems.
updated MenuWidgetStats to animate it's menuItems.

since Animate() scales sprites down to 1.0, we no longer need to scale menuItems from InventoryScreen.
removed that section of code from InventoryScreen.
playtested, working, with some minor bugs and alignment issues.