bug: boss decal welcome mat right doesn't flip properly. cause: from SetType() -> SetRotation() was setting the flipHorizontally boolean to false, by default. solution: this boolean is set false in ResetObject(), so there's no need to set it false again in SetRotation(). simply removed the line of code that set the boolean false in SetRotation(). however, this creates a bug with the sword projectile.
bug: sword projectile doesn't flip properly for Up and Right directions.
cause: a variety of issues, related to how the direction was being set via SpawnProjectile().
solution: a refactoring of SpawnProjectile(), SetRotation(), and SetType().
^ this was a very involved fix.

playtested, working as expected.
overall, very happy with the discovery and bugfixes.
it cleaned up a lot of smelly code, that was doing thing redundantly and needlessly.