created named MenuItem pointers in OverworldScreen to simplify and ease the setting of neighbors. updated LoadContent() to use these new named pointers, simplifying the naming and placing routines as well. various organization for OverworldScreen. setup the neighbors for named instances, which was actually very intuitive and easy.
integrated the movement routines for menuItem neighbors from InventoryScreen to OverworldScreen.
this is a direct integration, so there is no confusion as to how it's being done, it's done the same way.
playtested, working as expected.
but it's still hard to see the selectionBox as it moves around.
so i need to fix that.