added 'Shop' value to RoomType enumerator. updated BuildDungeon() to populate the dungeon.rooms list differently based on dungeon.type.
updated FinishRoom() to create a shop level/room based on room.type
for now, this room just has an exit object, without any mobs or door or decorations.
playtested, working, with a minor bug: theres a hole in the wall where the boss door was.. hmm..
but otherwise, it was super easy to add in the shop dungeon/room.