updated VendorScreen's constructor to accept a GameObj parameter. updated interactionFunctions.Interact() to pass the vendor obj to VendorScreen.
updated ForSaleWidget to default it's for sale menuItems to unknown.
updated VendorScreen to check for vendor type and set forSale widget's menuItems accordingly.

added a 'price' field to menuItem, as a byte.
updated VendorScreen.Input() to display the currently selected menuItem's price in the title of ForSale widget.

updated GameObjectFunctions.SetType() to set VendorAdvertisement objGroup to object, and texture to mainSheet.
this fixes an issue where an advertisement could of been passed to the VendorScreen.
playtested, working as expected.