updated the frame/animList values for: + reward particles + heart consumables + inventory gold + heart pieces + key & map + unknown menuItem + save/load menuItem + monitor, controller, sound menuItems + selection box sprite playtested, a bug.
bug: the loadout widget's heart pieces menuItem doesn't display correctly.
solution: updated the frame values being set in loadout widget to reflect the mainsheet repack.
playtested, working as expected.

fixed a minor non-bug that caused the stat menuItems to display incorrectly.
but, we don't really have stat menuItems right now, so just commented out the code for them.
this causes the stat menuItems to fall thru to unknown, which better reflects their state in the codebase at the moment.
playtested, working as expected.