changed ItemFunctions.HandleEffect() to UseItem(). moved ActorFunctions.UseWeapon() routines into ItemFunctions.UseItem() routines.
updated UseItem() to prevent statelock on an actor trying to use an empty bottle.
this fixes a bug where the hero would display 1 frame of the use animation when using an empty bottle.
playtested, working, above bug fixed.

updated InventoryScreen.SetLoadout() to allow player to set an empty bottle as the equipped item.
it's a bad idea to equip an empty bottle, but preventing player from doing so breaks consistency of inventory interaction.
playtested, working as expected.