added ProjectileBomb value to ObjType enumerator. updated GameObjAnimListManager with anim list for ProjectileBomb. added codepath for ProjectileBomb to GameObjFunctions.SetType().
added ItemBombs value to MenuItemType.
updated MenuItemFunctions.SetMenuItemData() with code path for ItemBombs.
updated the price of ItemBombs and ItemBombs.

created bombsCurrent and bombsMax bytes in SaveData struct.
updated constructor to start hero with 0 bombs, max of 99.
later we can limit the max, and use some upgrade to increase it, LttP style.

updated the ForSale widget to display ItemBomb and ItemBombs.
updated the InventoryWidget to display ItemBombs, if saveData.bombsCurrent is > than 0.
added a display amount to InventoryWidget to display the saveData.bombsCurrent amount.

playtested, bombs partially implemented, but theres MUCH more to do.