Update the Inventory screen to properly display the bow weapon, and the arrow count
updated InventoryScreen.SetLoadout() with path for WeaponBow.
this allows the player to equip the bow.
created InventoryWidget.SetInventoryMenuItems().
this makes it a bit easier to find, as it was buried in the widget's reset method.
organized codepaths in SetInventoryMenuItems(), added path for WeaponBow.

added arrowsDisplay ComponentAmountDisplay to InventoryWidget.
added paths to place, update, and draw the arrowsDisplay to InventoryWidget.
playtested, working as expected.