Simpifying projectile alignment, fixing arrow hitboxes.
refactor of Functions_Projectiles.SpawnProjectile().
refactor of Functions_GameObj.SetType().
updated Functions_GameObj.SetType() to set hitbox of arrow based on arrow's direction.

playtested, minor visual error with arrow sprite.
it's off by 1 pixel when facing different directions.
this causes the hitbox to appear to be misaligned.
however, the code to set the hitbox is the same for up/down and left/right.
so, it's an issue with the arrow sprite.

updated the arrow sprite to be 2 pixels, instead of 1 pixel wide.
playtested, collision recs aligning to arrow sprite perfectly for all directions.