Making Vendor products appear consistently, part 2
the ring disappears after purchase from the equipment vendor.
it really shouldn't do so, as this is inconsistent with how other items appear post-purchase.

updated the ForSale widget to always display the ring product from equipment vendor.
updated the VendorScreen to check saveData boolen for ring, call the AlreadyPurchased() dialog if true.
playtested, working as expected.