Fix the arrow sprite
reverted the arrow projetile sprite back to being a single pixel width.
it was just too wide for my preference, in the room.
the ui representation of the arrow is 3 pixels wide, and it looks fine.
the reason the arrow sprite was made 2 pixels wide was due to an alignment issue.
the hitbox isn't aligned visually to the sprite, it's wider by a pixel.
but, to make the arrow sprite look good, it needs to be 1 pixel wide.
it's' fine. playtested, looks better.