Fix confusion between projectiles, particles, pickups & gameobjects - part 2
the confusion between projectile, particles, and pickups continues with the FUnctions_Projectiles class.
this class of functions serves more than just projectiles, so it needs to be renamed.
plus, it's SpawnProjectile() method handles more than just projectiles, so that needs to be renamed as well.

renamed Functions_Projectiles to Functions_Entity.
renamed SpawnProjectile() to SpawnEntity().
updated various references to old class and it's methods.
playtested, working as expected.