Fairy death-rez polish part 1
if the hero has a fairy in a bottle, it brings him back to life upon death.
however, this currently happens immediately and doesn't even play the hero's death animation.
this isn't even close to the LttP implementation, the hero's death animation should complete.
then afterwards, the fairy animation should happen, so the player has time to understand what's going on.

organized & further simplified how DungeonScreen handles input.
playtested, working as expected.

modified how hero's death is handled in Functions_Actor.Update() & SetDeathState().
playtested, and working, but not quite as polished as i'd like it to be.
the death animation completes, but doesn't pause for a moment prior to fairy bottle animation.