Adding Room Builder/Editor - Part 5
we need to create the room builder widget, which holds the room objects and enemy spawns.
this widget's construction is based on the room builder mockup.

created WidgetRoomBuilder class, based on the other widgets created so far.
added WidgetRoomBuilder instance to global widget's list in Widgets class.
updated RoomBuilder screen to use the RoomBuilder widget.
removed the window reference RoomBuilder screen was using for testing.
playtested, working as expected.