Adding Room Builder/Editor - Part 7
we want to be able to set the 'active' object.
we want to be able to display which obj is the active object, using a selection box sprite.

added selection box component sprite to room builder widget.
updated activeObject to point to the first obj on the objList.
updated selection box to match the active object's position.
added path to pulse, scale, and draw the selectionBox.
playtested, working as expected.