Adding Room Builder/Editor - Part 8
we want to be able to click on an object, and set it as the active obj.
we also want to fix how the room builder widget is reset when room builder screen is loaded.

added the Reset() method back to RoomBuilder Widget.
except, we're not setting the position based on the X, Y parmaters.
we're resetting the window's UI instances, and resetting the activeObj & selectionBox.

created RoomBuilderWidget.SetActiveObj(int index) to handle setting the active object.
this will always point to an obj from the objList, thus the int index.
updated Reset() to use SetActiveObj().
added routine to check if user has clicked on any object from the room builder widget's objList.
this routine was added to RoomBuilderScreen.HandleInput().
playtested, working as expected.