Adding Room Builder/Editor - Part 12
we want to be able to set the currently selected tool.
we need another selection box for the current tool.
which means we need to rename the selection box to selectionBoxObj.
and create a new component sprite named selectionBoxTool.

created instance for activeTool pointer.
renamed selectionBox to selectionBoxObj, created selectionBoxTool instance.
updated Draw() to handle selectionBoxTool instance.
created BuilderWidget.SetActiveTool(), similar to SetActiveObj().
created BuilderWidget.UpdateSelectionBox() to handle updating both obj and tool selection boxes.
playtested, not working as expected...

bug: can't select tool icons.
cause: we're using the sprite component's drawRec for collision checking..
and that doesn't work, because it represents the sprite's texture region in the spritesheet.
we'll need to actually have a collision component associated with the sprite.
this collision component will need to be aligned to the sprite.
then we can check collisions against the collision component.

but, everything else works besides the collision checks with the tool icons.