Adding Room Builder/Editor - Part 15
we need to be able to select the toolbar icons now.
which means we simply need to check collisions between them and mouse position.
this is just like how we currently check collisions against room objs in widget.

updated RoomBuilderScreen.HandleInput() with paths to check for room objs & toolbar objs.
playtested, working as expected.