Adding Room Builder/Editor - Part 16
we need to add the actor objects, which will be gameobjects.
we'll set them up the same way we set up the toolbar ocns
we'll need to update various codepaths that rely on the current list arrangement.

added gameobject instances for 5 enemy actors, correctly placed in widget window.
updated RoomBuilderWidget.Reset().
updated RoomBuilderScreen.HandleInput() to work on roomObjs, actors, and toolbar objs.
playtested, working as expected.