Adding Room Builder/Editor - Part 18
we need to be able to interact with the buttons, just like the debug menu does.

created a List<ComponentButton> called buttons in RoomBuilderWidget.
switched to using this list instead of individual instances of component buttons.
this makes the button interaction consistent with how it's done in debug menu.
updated RoomBuilderScreen HandleInput() to check collisions with buttons.
playtested, working as expected.