Adding Room Builder/Editor - Part 20
created room instance in RoomBuilderScreen.
modified RoomBuilderScreen.Update() to track camera to center of room.
this is done in update so that if the room size changes, the camera automatically adapts - smoothly too.
updated RoomBuilderScreen.Draw() to use the camera view to draw the room.

added 'Dev' value to RoomType enum, for testing of RoomBuilder Screen.
updated BuildRoom() method to set a default value for floors being built.
playtested, working as expected... finally.
checked that the cursor collision rec aligns to the mouse.