Adding Room Builder/Editor - Part 30
when we build an exit room, the hero should not be placed at the exit
we should place the hero at the exit only when the dungeon is built

moved code that placed hero from Functions_Room.FinishExitRoom() to Functions_Dungeon.BuildDungeon().
changed hero placement code to reference dungeon.rooms[0].
organized BuildDungeon().
playtested, working as expected.