Adding Room Builder/Editor - Part 32
the door from the hub room to the boss room should be a boss door.

reordered what rooms are placed on the buildList.
the exit room is evaluated first, against the hub room.
this makes the boss door index 0 on the doorLocations list.
updated SetDoors() to check and see if room is hub room.
if so, and if the doorLocation index is 0, then the door is set to a boss door.

playtested - theres a bug with the Poke() method that is incorrectly placing doors.
this only happens in the Check Down section.
this section was never evaluated before, because we connected rooms in the Up direction.
but because the buildList has been reordered, we are now Poking rooms in the Down direction.

fixed an alignment issue with Poke(), in the Check Down section.
playtested, working as expected.