Adding Room Builder/Editor - Part 33
we want torch decorations around bossDoor, and boss decals in front of door.

updated BuildRoom() to SetDoors() prior to finishing a room.
updated FinishHubRoom() to locate and store a reference to the bossDoor in the room.
updated FinishHubRoom() to place torches on left and right of bossDoor ref.
updated FinishHubRoom() to place boss decals in front of bossDoor ref.
playtested, working as expected.