Adding Room Builder/Editor - Part 37
initial setup for saving room data.

created Functions_SaveLoad class.
added call to Functions_SaveLoad.Save() from RoomBuilder screen's HandleInput().
began working out the details of the room builder, only to realize that it should not be made in a UWP project.
UWP uses .net core, which isn't mature, and does things it's own way, which is frustrating.
so instead, we'll make the room builder using a directX project.
this requires a bit of planning..