Animate the LSN Screen's arrow
updated LSN screen to set arrow relative to action text.
updated how arrow text is placed so it animates left to right.
past a certain limit, the arrow is reset to align with action text.
this causes the arrow to 'bump' left to right, and move vertically with action text.
playtested, working as expected.