Notify Player after their game has been saved/created
refactored DialogScreen to use an ObjType enum, instead of a GameObj.
this allows us to pass the type as a constructor parameter..
instead of creating an obj, then setting it's type, then passing it in.
so, less code that's easier to understand.

minor refactoring of Functions_GameObject.SetType().
created Dialog enumerator to represent all possible dialog states/strings.
refactored DialogScreen to accept a Dialog enum as the only constructor parameter.
this sets both the speaker sprite/obj, and the dialog string.
updated LSN screen to pass Dialog.GameCreated or Dialog.GameSaved to DialogScreen.

playtested that guide displays after save/new, working as expected.