Polish loading case where game file is missing
added GameNotFound value to Dialog enumerator.
updated Functions_Backend.LoadGame() to display a dialog screen, if file to be loaded doesn't exist.
updated DialogScreen with paths to handle GameNotFound value.
if file to be loaded doesn't exist, the current game data is saved to that game slot.
playtested, working as expected.