Polish loading case where game file is corrupted
added GameLoadFailed value to Dialog enumerator.
updated Functions_Backend.LoadPlayerData() to handle malformed xml.
playtested loading malformed xml, working, but not as expected.

the xml header has to be malformed for the catch{} to be called.
otherwise the game loads any data it can from the xml.
this can cause a state where a save file is only partially loaded.
not really sure how to check to make sure the xml is perfectly formatted.
but, it works for now.