Polish LSN Screen MenuItem Interaction
when player selects a game file on LSN screen, the bkg menuItem scales up to 2.0.
but a dialog screen is immediately created, leaving the menuItem scale at 2.0.
this looks bad, so don't scale up the menuItems on LSN screen anymore.

updated LSN screen to remove scaling and animation of bkg menuItems.
playtested, working as expected.