Fix Corrupted Autosave File Bug
upon a continue game, the corrupted save file dialog displays.

cause: when we continue a game, the autosave is loaded.
but now we are creating a shop screen, which creates a shop dungeon.
upon dungeon creation, the autosave file gets written.
so we have a conflict where autosave file is being loaded and saved to at the same time.

fix: dont load and save the autosave file at the same time.
updated Functions_Dungeon.BuildDungeon() to NOT autosave upon dungeon creation.
updated DungeonScreen.Update - Closed state to autosave.
now whenever the player leaves the dungeon screen, autosave will be called.

playtested, fixed a minor bug where shop dungeon was being created twice.
this involved an update to the TitleScreen. now working as expected.