Polish the DialogScreen
added a ScreenRec background to DialogScreen.
updated DialogScreen to fade black background in when screen opens.
updated DialogScreen to fade black overlay in when screen closes.
this creates smooth transitions between DialogScreen and most other screens.
the exception is when the player 'talks' with the guide in the shop.
playtested, working as expected.