Polish Dialog Screen User Experience
when the player talks with the guide in the shop, dont fade DialogScreen's background or overlay rec in.
this is done by checking the dialogType in dialogScreen.

added fadeBkg boolean to dialogScreen.
updated dialogScreen to only fade bkg & overlay in if fadeBkg is true.
fadeBkg is based on the dialogType,
fadeBkg is set (along with the dialogString) in dialogScreen.LoadContent().
playtested, working as expected, but needs additional polish for save game case.