Refactor how the DialogScreen updates DungeonScreen
when the player talks to the guide, or opens the dialog screen in any way..
the dialog screen updates the dungeon screen, if the dungeon screen exists.
but, this is not the desired behavior.
instead the dungeon screen should only update if the player is talking to the guide.

added boolean updateDungeonScreen to DialogScreen.
by default this boolean is set to true.
then, based on the dialog type, the boolean is updated.
this lets the load/save/new game dialogs to set the boolean.
playtested, working as expected.