Fix Audio bug that causes clipping
Because the same sound effect can be called many times, at the same time..
we use a list of sound effect instances, which are looped thru..
and if the game finds a sound not being played on the list, it gets played.
and any other playing other sound effect instances on the list are stopped.
when they are stopped, then causes clipping in the audio.
they shouldn't be stopped, they should be allowed to finish.

updated Assets.Play() to not stop a playing sound effect instance.
playtested, working as expected. some sfx blends together.
but this is better than the clipping that was happening prior.