Fix Bug where LSN Screen does not update after a game has been saved
we should be displaying the latest data from the saved file
when you save a game, the LSN screen does not update the current save slot

warning - this is a complicated update.
in order to make this work properly, alot of stuff had to be changed.
there was alot of trial and error with this commit, but it works now.
the biggest hack here is that the game is reloading all game files,
from the LSN Screen, when Screen.Unload() is called.
this is done for complicated reasons involving the creation of a new game.

the LSN screen needs the latest game files loaded, in order to display them correctly.
however, this loading is async, which means we never know when they are loaded.
initially, we load all the game files during the creation of the TitleScreen.
in addition, we reload the files after the LSN screen closes, in prep for it's next opening.
this gives the game enough time to finish loading before the LSN screen opens.

updated LoadGame() to better manage where PlayerData.current points to.
updated LSN Screen with UpdateCurrentDisplay(), which is called after player saves a game.
the dialog screen was also updated to prevent user input / exiting of screen,
until the dialog text is finished being displayed on screen.

lots of playtesting, and i think it's working as expected.
honestly, there needs to be more playtesting done.