Only process add/move/delete if user clicks within room
we don't want the user adding objects outside of the room, only inside.
same with deleting and moving objects.
if an objet is outside of the room.rec (walls for example), the obj cannot be modified.
this prevents the user from deleting or moving walls, which is good.

updated RBscreen.HandleInput() to check if cursor is contained within currentRoom.
if true, then add/move/delete is processed normally.
playtested, user is limited to clicking within the room as intended.