In RoomBuilder, allow User to select a room xml file to load
updated LoadRoomData(), with string parameter for fileaddress.
created SelectRoomFile(), which presents user with a file dialog.
updated both functions to modify a passed RoomXmlData parameter.

created RMScreen.BuildFromRoomData().
updated SelectRoomFile() to accept a RoomBuilderScreen parameter.
this allows the function to be async, and call BuildFromRoomData() when ready.
playtested, working as expected.