Fix door decoration placement
a left/right door's decorations are placed left and right of door.
they should be placed top and bottom of door.

updated Functions_Room.DecorateDoor() to check the Door.direction.
based on this direction, door decoration positions are set.
this allows decorations to be placed left/right, and top/bottom around doors.
playtested, working as expected.