Add Flag Boolean to control creation of Interaction Dialogs
added Flags.ShowDialogs boolean.
this controls if dialog screens are created when hero interacts with an object.
updated hero -> chest interaction with checks to Flags.ShowDialogs.
updated hero -> boss door interaction with check to Flags.ShowDialogs.
playtested, working as expected.

turning off dialogs will be useful when players are doing speedruns.
we can add this to the game options screen later on too.