Display the amount of arrows and bombs Hero has in WorldUI
updated Functions_WorldUI.Update() to set the visibility of weapon & item amounts.
this is based on the weapon and item that hero has equipped.
the amount display is visible for the weapon: bow, and the item: bombs.
each amount display's value is set based on the weapon or item equipped.
playtested, working beautifully.