Animate the dungeon map's scroll open and closed
updated DungeonMap to animate right scroll to right on opening.
updated Dungeonmap to animate left scroll to left on closing.
updated DungeonMap to draw bkg sprites based on their position.
this position is compared to the left/right scrolls based on display state.
this makes the scroll appear to open and close.
playtested, looking really good and working as expected.