Update Overworld Screen to use a Scroll background
updated Overworld Screen to use Scroll data class & Functions_Scroll.
this replaces the window instance that OverworldScreen was using.
since this is a map, it should be consistent with the other map (dungeon map).
playtested, working as expected.
this was done in prep for the new overworld map that's in development rn.