Simplify Dungeon Generation Routine
because we are checking room.type in Poke()..
we no longer require the boss room to be at index 0 of the build list.
nor do is it necessary to have hub room as index 1 of the build list.
removed codepath that set buildList index 0 and index 1.
playtested, working as expected.