Improve Bomb Projectiles
fixed a bug where bumper's set an object's speed to 0.
this caused the object to 'slide' to stillness after bumper collision.

changed the Offset for bomb projectile when actor throws a bomb.
updated InteractObject() to StopMovement() on a bomb that collides with a blocking object.
increased the initial push / velocity of a bomb being thrown.
updated KillProjectileUponCollision() to kill bombs as well.
this means bombs explode when spikeblocks hit them, which is intended.
lots of playtesting, working as expected.