Fix Arrow & Bombs Bug
because we changed how collisions + interactions were handled,
we introduced two small bugs related to arrows and bombs.
when the hero shoots arrows left or right, they collide with the hero.
when the hero throws a bomb south, it collides with the hero and stops moving.

updated Functions_Alignment.SetOffsets() to properly place arrows and bombs.
playtested, bugs fixed.