Fix Pit COllision Bug
if two actors collide, they are prevented from moving through each other.
this is the foundation of the collision system for actors.
however, when two actors are being pulled into a pit at the same time,
they collide, and their movement into the pit is canceled out.
this leads to a state where two actors hover over the pit, never falling in,
stuck in a hit state where they are unmove-able. this is bad.

so, the fix is to update the actor colliding with the pit,
in a manner that does not check collisions. ignore collisions, basically.
but since this interaction is happening from a collision check,
we can't just projectMovement() and then align()...
because the parent collision check will revert the actor back to
the compMove.position and ignore the compMove.newPosition.

the fix is to set the compMove.position directly using the actor's magnitude.
then set the compMove.newPosition to compMove.position, then Align() components.
this force-moves the actor into the pit, ignoring any blocking collisions.
updated Interact() -> Pit routine with force-move code for actors.
playtested, working as expected.