Fix Sword Hit Particle Bug
when a sword projectile hits an actor, it spawns a hit particle at the swords tip.
however, this was being done each frame, so the sword spawned about 100 particles.
the fix is to check the sword.lifeCounter value, making sure the lifeCounter = 1.
this limits the hit sparkle spawn to the first frame of sword's existence.
this is the same approach used when checking sword collisions with blocking roomObjs.
this fixes the creation of 100s of particles upon a sword collision with actor.
playtested, working.

but, this sword interaction still creates 2 sparkles.
this is because the projectile's collision system checks collisions per axis.
so, i modified CheckCollisions(Entity) to perform a single projection collision check.
this prevents projectiles from interacting with actors twice per frame.
this also saves some ticks per frame.
playtested, working as expected.